requirements.txt
asgiref==3.2.7
Django~=2.2.12
django-braces==1.14.0
django-countries==6.1.2
django-credit-cards==0.4.1
django-crispy-forms==1.9.0
django-model-utils==4.0.0
django-star-ratings==0.9.1
Pillow==7.1.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
swapper==1.1.2.post1
django-oscar~=2.0.4